# src/integration-test.disabled/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: rtj-management
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:59560
  description: Generated server url
...
paths:
...
  /{municipalityId}/{namespace}/errands:
    get:
      tags:
      - Errands
      summary: Find errands
      description: Query for errands with optional spring-filter expression
      operationId: findErrands
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        description: Namespace
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: filter
        in: query
        description: "Syntax description: [spring-filter](https://github.com/turkraft/spring-filter/blob/85730f950a5f8623159cc0eb4d737555f9382bb7/README.md#syntax)"
        required: false
        schema:
          type: string
        example: status:'NEW' and created>'2024-01-01T00:00:00.000+01:00'
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "touched,DESC"
          items:
            type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindErrandsResponse"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
      - Errands
      summary: Create errand
      description: Creates a new errand
      operationId: createErrand
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        description: Namespace
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Errand"
        required: true
      responses:
        "201":
          description: Successful operation
          headers:
            Location:
              style: simple
              schema:
                type: string
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/{namespace}/errands/{errandId}:
    get:
      tags:
      - Errands
      summary: Read errand
      description: Fetches the errand that matches the provided id
      operationId: readErrand
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        description: Namespace
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: errandId
        in: path
        description: Errand id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errand"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
      - Errands
      summary: Delete errand
      description: Deletes the errand
      operationId: deleteErrand
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        description: Namespace
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: errandId
        in: path
        description: Errand id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Successful operation
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    patch:
      tags:
      - Errands
      summary: Update errand
      description: Patches the errand with the supplied attributes
      operationId: updateErrand
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        description: Namespace
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: errandId
        in: path
        description: Errand id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchErrand"
        required: true
      responses:
        "204":
          description: Successful operation
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        detail:
          type: string
        title:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        detail:
          type: string
        causeAsProblem:
          $ref: "#/components/schemas/ThrowableProblem"
        instance:
          type: string
          format: uri
    ThrowableProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        causeAsProblem: {}
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    Errand:
      type: object
      description: Errand envelope
      properties:
        id:
          type: string
          description: Unique identifier of the errand
          examples:
          - cb20c51f-fcf3-42c0-b613-de563634a8ec
          readOnly: true
        municipalityId:
          type: string
          description: Municipality id
          examples:
          - "2281"
          readOnly: true
        namespace:
          type: string
          description: Namespace
          examples:
          - MY_NAMESPACE
          readOnly: true
        errandNumber:
          type: string
          description: Human-readable errand number
          examples:
          - CAREM-2026-00042
          readOnly: true
        typeSlug:
          type: string
          description: Registered errand type slug — validated against ErrandTypeRegistry
          examples:
          - case-type-slug
        title:
          type: string
          description: Title for the errand
          examples:
          - Title of the errand
        status:
          type: string
          description: Status of the errand
          examples:
          - NEW
        description:
          type: string
          description: Description of the errand
          examples:
          - Long description text
        priority:
          type: string
          description: Priority of the errand
          examples:
          - HIGH
        reporterUserId:
          type: string
          description: User id of the reporter
          examples:
          - joe01doe
        assignedUserId:
          type: string
          description: User id of the assignee
          examples:
          - jane02doe
        processDefinitionName:
          type: string
          description: Name of the Operaton process definition to start when the errand
            is created
          examples:
          - Handläggning av ärende
        processInstanceId:
          type: string
          description: Id of the Operaton process instance started for this errand
          examples:
          - a-process-instance-id
          readOnly: true
        created:
          type: string
          format: date-time
          description: Created timestamp
          readOnly: true
        modified:
          type: string
          format: date-time
          description: Modified timestamp
          readOnly: true
        touched:
          type: string
          format: date-time
          description: Touched timestamp
          readOnly: true
...
    PatchErrand:
      type: object
      description: PatchErrand model — patchable envelope fields only
      properties:
        title:
          type: string
          description: Title for the errand
          examples:
          - Title of the errand
        status:
          type: string
          description: Status of the errand
          examples:
          - NEW
        description:
          type: string
          description: Description of the errand
        priority:
          type: string
          description: Priority of the errand
          examples:
          - HIGH
        reporterUserId:
          type: string
          description: User id of the reporter
          examples:
          - joe01doe
        assignedUserId:
          type: string
          description: User id of the assignee
          examples:
          - jane02doe
    Direction:
      type: string
      enum:
      - ASC
      - DESC
    FindErrandsResponse:
      type: object
      description: Paged errand response
      properties:
        errands:
          type: array
          items:
            $ref: "#/components/schemas/Errand"
            readOnly: true
        _meta:
          $ref: "#/components/schemas/PagingAndSortingMetaData"
          readOnly: true
    PagingAndSortingMetaData:
      type: object
      description: PagingAndSortingMetaData model
      properties:
        page:
          type: integer
          format: int32
          description: Current page
          examples:
          - "5"
          readOnly: true
        limit:
          type: integer
          format: int32
          description: Displayed objects per page
          examples:
          - "20"
          readOnly: true
        count:
          type: integer
          format: int32
          description: Displayed objects on current page
          examples:
          - "13"
          readOnly: true
        totalRecords:
          type: integer
          format: int64
          description: Total amount of hits based on provided search parameters
          examples:
          - "98"
          readOnly: true
        totalPages:
          type: integer
          format: int32
          description: Total amount of pages based on provided search parameters
          examples:
          - "23"
          readOnly: true
        sortBy:
          type: array
          items:
            type: string
            description: The properties to sort by
            examples:
            - property
            readOnly: true
        sortDirection:
          $ref: "#/components/schemas/Direction"
          description: The sort order direction
          examples:
          - ASC
...
  securitySchemes: {}